testops_copilot/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==1.6.5

# Celery и очереди
celery==5.3.4
redis==5.0.1
redis[asyncio]==5.0.1
flower==2.0.1

# LangGraph и агенты
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# HTTP клиенты
httpx==0.25.2
aiohttp==3.9.1

# OpenAI клиент для Cloud.ru Foundation Models API
openai==1.12.0

# Валидация и парсинг
pydantic==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1
jinja2==3.1.2
openapi-spec-validator==0.7.1

# Тестирование и валидация
pytest==7.4.3
pytest-asyncio==0.21.1
allure-pytest==2.13.2
bandit==1.7.5

# Playwright для UI анализа
playwright==1.40.0

# ML и embedding
scikit-learn==1.3.2
numpy==1.26.2

# Утилиты
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# GitLab интеграция
python-gitlab==4.2.0

# GitHub интеграция
PyGithub==2.1.1

# Логирование и мониторинг
structlog==23.2.0
prometheus-client==0.19.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

